PROJET/requirements.txt
google-adk==1.26.0
python-dotenv==1.2.2
litellm>=1.75.5
